PR Review: Reset custom form on configuration page, made onResetWaymarkInstance an AppProvider function to pass down, cleaned up styling
